handle wishlist in most comp, add comfirm message Dialoge service, handle notification

+++ frontend/src/app/Components/wishlist/wishlist.component.ts
@@ -1,69 +1,75 @@
 import { CommonModule } from '@angular/common';
 import { Component, OnInit } from '@angular/core';
-import { AuthService } from '../../services/auth/auth.service';
+import { MatDialog } from '@angular/material/dialog';
+import { MatSnackBar } from '@angular/material/snack-bar';
 import { RouterModule } from '@angular/router';
-import { ProductService } from '../../services/product.service';
 import { CartsService } from '../../services/carts.service';
 import { WishlistService } from '../../services/wishlist/wishlist.service';
-import { MatSnackBar } from '@angular/material/snack-bar';
+import { ConfirmDialogComponent } from '../confirm-dialog/confirm-dialog.component';
+// Adjust path
 
 @Component({
   selector: 'app-wishlist',
   standalone: true,
-  imports: [CommonModule,RouterModule],
+  imports: [CommonModule, RouterModule],
   templateUrl: './wishlist.component.html',
-  styleUrl: './wishlist.component.css'
+  styleUrls: ['./wishlist.component.css']
 })
-export class WishlistComponent implements OnInit  {
-  isLoggedIn!:boolean
+export class WishlistComponent implements OnInit {
   wishlist: any[] = [];
 
   constructor(
     private notification: MatSnackBar,
-    private wishlistService : WishlistService,
-    private cartServive : CartsService){}
-
-
+    private wishlistService: WishlistService,
+    private cartService: CartsService,
+    private dialog: MatDialog
+  ) {}
 
   ngOnInit(): void {
-     this.wishlistService.getWishList().subscribe(state => this.wishlist = state) //fill wishlist
-
-   }
-
-//remove from wishlist
-  removeFromWishlist(id: number) {
-    this.wishlistService.removeFromWishList(id);
-    this.notification.open("Item Removed successfully!", "Close", {
-      duration :1000,
-      horizontalPosition: 'right',
-      verticalPosition: "top",
-      panelClass: 'alert-red'
-    })
+    this.wishlistService.getWishList().subscribe(state => this.wishlist = state);
   }
-//move to cart
+
+  removeFromWishlist(id: number) {
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      width: '300px',
+      data: { message: 'Do you really want to remove this item from your wishlist?' }
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result === 'confirm') {
+        this.wishlistService.removeFromWishList(id);
+        this.notification.open("Product removed from wishlist successfully!", "Close", {
+          duration: 1500,
+          horizontalPosition: 'right',
+          verticalPosition: 'bottom',
+          panelClass: 'alert-red'
+        });
+      }
+    });
+  }
+
   moveToCart(item: any) {
-    if(!this.isProdInCart(item.id)){ 
-      this.cartServive.addProductToCart(item)
-      this.notification.open('Item added to cart successfully!', 'Close', {
+    if (!this.isProdInCart(item.id)) {
+      this.cartService.addProductToCart(item);
+      this.notification.open('Product added to cart successfully!', 'Close', {
         duration: 1000,
         horizontalPosition: 'right',
-        verticalPosition: 'top',
+        verticalPosition: "bottom",
         panelClass: 'alert-red'
-      })
-    }else{
-      this.cartServive.removeProductFromCart(item.id) 
-      this.notification.open("Item Removed from cart successfully!", "Close", {
-        duration :1000,
+      });
+    } else {
+      this.cartService.removeProductFromCart(item.id);
+      this.notification.open("Product removed from cart successfully!", "Close", {
+        duration: 1000,
         horizontalPosition: 'right',
-        verticalPosition: "top",
+        verticalPosition: "bottom",
         panelClass: 'alert-red'
-      })
+      });
     }
-    
   }
 
-//is in Cart
-  isProdInCart(id:number): boolean {
-    return this.cartServive.isProductInCart(id) 
+  isProdInCart(id: number): boolean {
+    return this.cartService.isProductInCart(id);
   }
+
 }
